feat: JSON import/export
Add the possibility to save and load COAs from the disk.

For individual COAs, this provides an alternative in the case
where the URL string is too long to be handled by the browser.

For a whole gallery, this gives a way to regroup a collection
of COAs into a single file.