QoL improvements (#185)
* feat: make uploaded charges persistent

* Rework shield and zoom options

- Shield and zoom can now be set on a COA-basis
- Use global shield option if the COA does not define one
- Remove global zoom option

* feat: allow naming COAs

* feat: JSON import/export

Add the possibility to save and load COAs from the disk.

For individual COAs, this provides an alternative in the case
where the URL string is too long to be handled by the browser.

For a whole gallery, this gives a way to regroup a collection
of COAs into a single file.